refactor: moved components to 'resources' folder
