+++ components/resources/NewResource.js
