+++ pages/learningUnits/[learningUnitId].js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useRouter } from "next/router";
-import NewResource from "../../components/NewResource";
+import NewResource from "../../components/resources/NewResource";
 
 export default function LearningUnits() {
   const router = useRouter();
